added functionality to DAL

+++ DomainModel/Models/Car.cs
@@ -14,7 +14,7 @@ namespace DomainModel.Models
         public string Model { get; set; }
         public int Front_Brake_Remaining { get; set; }
         public int Rear_Brake_Remaining { get; set; }
-        public int Anti_Freeze_Strength { get; set; }
+        public decimal Anti_Freeze_Strength { get; set; }
     }
 
 }
